430

Former-commit-id: 41cba7b9d7c10a0f521bb0522b9720ae5ffe2236 [formerly 7697d449a015ebc2c9648c11feb8defacad28660 [formerly b49b063778c6987c0f46876181e2dde22c8a2b5e]]
Former-commit-id: 7697d449a015ebc2c9648c11feb8defacad28660
Former-commit-id: 6c4359bfc54316a39a6cdb98bce3b266c8fe8c55

+++ gdsfactory/components/ring_single_heater.py
@@ -20,7 +20,7 @@ def ring_single_heater(
     cross_section_heater: CrossSectionFactory = gf.cross_section.strip_heater_metal,
     cross_section: CrossSectionFactory = gf.cross_section.strip,
     contact: ComponentFactory = contact_heater_m3_mini,
-    port_orientation: int = 90,
+    port_orientation: float = 90,
     contact_offset: Float2 = (0, 0),
     **kwargs
 ) -> gf.Component:
